# backend/api/docs/swagger.yaml
basePath: /
definitions:
...
  internal_handler_chat.ChatRequest:
    properties:
      config:
        additionalProperties: true
        type: object
      features:
        items:
          type: string
        type: array
      history:
        items:
          $ref: '#/definitions/internal_handler_chat.HistoryMessage'
        type: array
      message:
        type: string
      missionId:
        type: string
      mode:
        type: string
      model:
        type: string
      sessionId:
        type: string
      skills:
        items:
          type: string
        type: array
      strategyVersion:
        type: string
    type: object
  internal_handler_chat.HistoryMessage:
    properties:
      content:
        type: string
      role:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@echo.app
    name: Echo API Support
    url: https://echo.app/support
  description: Core API Server for Echo Platform providing Auth, Chat, Sessions, Models,
    Settings, Admin, Memory, and System Services.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://echo.app/terms
  title: Echo Backend API
  version: "1.0"
paths:
...
  /api/v1/chat:
    post:
      consumes:
      - application/json
      description: Forwards a user message to the agent for processing
      parameters:
      - description: Chat payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler_chat.ChatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Send a chat message
      tags:
      - Chat
...
  /api/v1/missions/{id}/approve:
    post:
      description: Approves a human-in-the-loop tool approval request for a mission
      parameters:
      - description: Mission ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Approve a pending tool call
      tags:
      - Chat
  /api/v1/missions/{id}/deny:
    post:
      description: Denies a human-in-the-loop tool approval request for a mission
      parameters:
      - description: Mission ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Deny a pending tool call
      tags:
      - Chat
  /api/v1/missions/{missionId}/stream:
    get:
      description: Streams real-time mission execution logs as Server-Sent Events
      parameters:
      - description: Mission ID
        in: path
        name: missionId
        required: true
        type: string
      produces:
      - text/event-stream
      responses:
        "200":
          description: Event stream
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Stream mission logs
      tags:
      - Chat
...
  /api/v1/skills:
    get:
      description: Returns the catalog of skills available to agents
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List available agent skills
      tags:
      - Chat
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer token authorization header (Format: Bearer <JWT>).'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"